+++ grpc-node/context.ts
@@ -2,8 +2,6 @@ import type {
   UntypedServiceImplementation,
   UntypedHandleCall,
   handleUnaryCall,
-  ServerUnaryCall,
-  sendUnaryData,
 } from "@grpc/grpc-js";
 import type { ServiceImplementationWrapper } from "@saflib/grpc-node";
 import {
@@ -13,13 +11,12 @@ import {
   type SafReporters,
   safReportersStorage,
   getServiceName,
-  getSafReporters,
 } from "@saflib/node";
 import { SafAuth } from "@saflib/grpc-specs";
 import { createLogger } from "@saflib/node";
-import { status } from "@grpc/grpc-js";
 import { AsyncLocalStorage } from "async_hooks";
 import { defaultErrorReporter } from "../node/src/errors.ts";
+import { runGrpcMethod } from "./runner.ts";
 
 export type SafServiceImplementationWrapper = (
   impl: UntypedServiceImplementation,
@@ -76,29 +73,6 @@ export const addSafContext: SafServiceImplementationWrapper = (
   }
   return wrappedService;
 };
-
-export function runGrpcMethod(
-  methodImpl: handleUnaryCall<any, any>,
-  call: ServerUnaryCall<any, any>,
-  callback: sendUnaryData<any>,
-) {
-  const { logError } = getSafReporters();
-  try {
-    const result = methodImpl(call, callback) as any;
-    if (result instanceof Promise) {
-      return result.catch((error) => {
-        const e = error as Error;
-        logError(e);
-        callback({ code: status.INTERNAL, message: e.message } as any, null);
-      });
-    }
-    return result;
-  } catch (error) {
-    const e = error as Error;
-    logError(e);
-    callback({ code: status.INTERNAL, message: e.message } as any, null);
-  }
-}
 
 export function makeGrpcServerContextWrapper(
   storage: AsyncLocalStorage<any>,
